Move the logic to modify default value for consuming segment inside RealtimeTableDataManager (#10086)
